tfc_web/requirements.txt
Django>=3.2,<4.0
python-dateutil
django-allauth
requests
whitenoise
psycopg2-binary
gunicorn
Pillow
pytz
markdown2
wheel
djangorestframework
coreapi
pygments
markdown
django-cors-headers
iso8601
zeep
django-filter
pyofo
pycryptodome
geopy
ipython
pandas
xlrd
untangle
pyproj
isodate
